Redo dllinjection
