+++ external/source/exploits/CVE-2018-8440/ALPC-TaskSched-LPE/ClassDiagram.cd
@@ -1,0 +1,2 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ClassDiagram />
