+++ external/source/exploits/CVE-2018-8440/ALPC-TaskSched-LPE/Resource.rc
