+++ external/source/exploits/CVE-2018-8440/common/ReflectiveDLLInjection.h
