+++ external/source/exploits/CVE-2018-8440/dll/reflective_dll.vcxproj
@@ -26,6 +26,22 @@
       <Platform>x64</Platform>
     </ProjectConfiguration>
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="src\exploit.h" />
+    <ClInclude Include="src\ntimports.h" />
+    <ClInclude Include="src\ReflectiveLoader.h" />
+    <ClInclude Include="src\resource.h" />
+    <ClInclude Include="src\rpc_h.h" />
+    <ClInclude Include="src\stdafx.h" />
+    <ClInclude Include="src\typed_buffer.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="src\Exploit.cpp" />
+    <ClCompile Include="src\Hardlink.cpp" />
+    <ClCompile Include="src\ReflectiveDll.c" />
+    <ClCompile Include="src\ReflectiveLoader.c" />
+    <ClCompile Include="src\rpc_c.c" />
+  </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{3A371EBD-EEE1-4B2A-88B9-93E7BABE0949}</ProjectGuid>
     <RootNamespace>reflective_dll</RootNamespace>
@@ -112,9 +128,10 @@
     <LinkIncremental>false</LinkIncremental>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <OutDir>$(SolutionDir)$(Platform)\$(Configuration)\</OutDir>
+    <OutDir>..\..\..\..\..\data\exploits\CVE-2018-8440</OutDir>
     <IntDir>$(Platform)\$(Configuration)\</IntDir>
     <LinkIncremental>false</LinkIncremental>
+    <TargetName>ALPC_TaskSched</TargetName>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
@@ -246,14 +263,6 @@
       <TargetMachine>MachineX64</TargetMachine>
     </Link>
   </ItemDefinitionGroup>
-  <ItemGroup>
-    <ClCompile Include="src\ReflectiveDll.c" />
-    <ClCompile Include="src\ReflectiveLoader.c" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="..\common\ReflectiveDLLInjection.h" />
-    <ClInclude Include="src\ReflectiveLoader.h" />
-  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
