+++ external/source/exploits/CVE-2018-8440/dll/src/ReflectiveLoader.c
@@ -37,7 +37,6 @@ HINSTANCE hAppInstance = NULL;
 // available (and no inline asm available under x64).
 __declspec(noinline) ULONG_PTR caller( VOID ) { return (ULONG_PTR)_ReturnAddress(); }
 //===============================================================================================//
-
 #ifdef ENABLE_OUTPUTDEBUGSTRING
 #define OUTPUTDBG(str) pOutputDebug((LPCSTR)str)
 #else /* ENABLE_OUTPUTDEBUGSTRING */
