+++ external/source/exploits/CVE-2018-8440/dll/src/ReflectiveLoader.h
