+++ external/source/exploits/CVE-2018-8440/dll/src/exploit.h
@@ -1,0 +1,7 @@
+#ifndef EXPLOIT_H
+#define EXPLOIT_H
+
+extern "C" __declspec (dllexport) void testFunc(void);
+//extern "C" __declspec (dllexport) void RunIt(void*);
+
+#endif
