+++ external/source/exploits/CVE-2018-8440/inject/src/GetProcAddressR.h
