--- external/source/exploits/CVE-2018-8440/rdi/ALPC-TaskSched-LPE/Header.h
@@ -1,3 +1,0 @@
-#include <Windows.h>
-
-extern "C" __declspec (dllexport) DWORD CALLBACK ExploitThread(LPVOID);
