--- external/source/exploits/CVE-2018-8440/rdi/ALPC-TaskSched-LPE/dllmain.cpp
@@ -1,24 +1,0 @@
-// dllmain.cpp : Defines the entry point for the DLL application.
-#include "stdafx.h"
-#include "Header.h"
-
-//extern "C" __declspec (dllexport) DWORD CALLBACK ExploitThread(LPVOID hModule);
-
-BOOL APIENTRY DllMain(HMODULE hModule,
-	DWORD  ul_reason_for_call,
-	LPVOID lpReserved
-)
-{
-	switch (ul_reason_for_call)
-	{
-	case DLL_PROCESS_ATTACH:
-		//CreateThread(NULL, 0, ExploitThread, hModule, 0, NULL);
-		break;
-	case DLL_THREAD_ATTACH:
-	case DLL_THREAD_DETACH:
-	case DLL_PROCESS_DETACH:
-		break;
-	}
-	return TRUE;
-}
-
