--- external/source/exploits/CVE-2018-8440/rdi/ALPC-TaskSched-LPE/stdafx.h
@@ -1,15 +1,0 @@
-// stdafx.h : include file for standard system include files,
-// or project specific include files that are used frequently, but
-// are changed infrequently
-//
-
-#pragma once
-
-
-
-#define RPC_USE_NATIVE_WCHAR
-
-#include <stdio.h>
-#include <tchar.h>
-#include <Windows.h>
-#include <memory>
